[TS #132] canEdit, canDelete Organization and User. Implementing delete button and such.

+++ frontend/src/app/components/event/event.component.ts
@@ -143,11 +143,11 @@ export class EventComponent implements OnInit {
   }
 
   private deleteEvent() {
-    this.eventService.deleteEvent(this.calendarEvent).subscribe(() => {
-      console.log('Event deleted');
-    }, error => {
-      // display Error.
-    });
+    if (confirm(`You are deleting "${this.calendarEvent.name}". Are you sure?`)) {
+      this.eventService.deleteEvent(this.calendarEvent).subscribe(() => {
+        console.log('Event deleted');
+      });
+    }
   }
 
   private getParticipants() {
